Remove CommunityStakeStep and related components, consolidating functionality into CommunityTransactions for improved structure and maintainability.
