--- packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityStakeStep/index.ts
@@ -1,3 +1,0 @@
-import CommunityStakeStep from './CommunityStakeStep';
-
-export default CommunityStakeStep;
