+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/EnableStake/validations.ts
