+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/SignStakeTransactions/SignStakeTransactions.tsx
