+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/SignStakeTransactions/useLaunchCommunityStake.tsx
@@ -8,8 +8,11 @@ import {
 import { useUpdateCommunityStake } from 'state/api/communityStake';
 import useUserStore from 'state/ui/user';
 import useAppStatus from '../../../../../../hooks/useAppStatus';
-import { useNamespaceFactory } from '../../CommunityTransactions ';
-import { ActionState, defaultActionState } from '../types';
+import { useNamespaceFactory } from '../../CommunityTransactions';
+import {
+  ActionState,
+  defaultActionState,
+} from '../../CommunityTransactions/types';
 
 interface UseLaunchCommunityStakeProps {
   namespace: string;
